GenericSpec: Ensure the MockingContainer is always initialized
It is somewhat common for spec writers to override BeforeEachExample, but forget
to call base.BeforeEachExample. If the spec being written inherits from
GenericSpec, this results in a somewhat obscure exception.

This refactors how GenericSpec plugs into Spec's Before pipeline such that the
initialization always happens, regardless of what happens in BeforeEachExample.